Move to lib/middleware/receive-message
