+++ lib/middleware/receive-message/draft-completed.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../helpers');
-const replies = require('../replies');
+const helpers = require('../../helpers');
+const replies = require('../../replies');
 
 module.exports = function postDraftSubmission() {
   return (req, res) => {
