+++ lib/middleware/receive-message/draft-photo.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../helpers');
-const replies = require('../replies');
+const helpers = require('../../helpers');
+const replies = require('../../replies');
 
 module.exports = function draftSubmissionPhoto() {
   return (req, res, next) => {
