+++ lib/middleware/receive-message/draft-why-participated.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../helpers');
-const replies = require('../replies');
+const helpers = require('../../helpers');
+const replies = require('../../replies');
 
 module.exports = function draftSubmissionWhyParticipated() {
   return (req, res, next) => {
