+++ lib/middleware/receive-message/menu-doing.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const replies = require('../replies');
+const replies = require('../../replies');
 
 module.exports = function doingMenu() {
   return (req, res, next) => {
