+++ lib/middleware/receive-message/phoenix-campaign-get.js
@@ -3,11 +3,11 @@
 const newrelic = require('newrelic');
 const logger = require('winston');
 
-const helpers = require('../helpers');
-const phoenix = require('../phoenix');
-const stathat = require('../stathat');
-const ClosedCampaignError = require('../../app/exceptions/ClosedCampaignError');
-const replies = require('../replies');
+const helpers = require('../../helpers');
+const phoenix = require('../../phoenix');
+const stathat = require('../../stathat');
+const ClosedCampaignError = require('../../../app/exceptions/ClosedCampaignError');
+const replies = require('../../replies');
 
 module.exports = function getCampaign() {
   return (req, res, next) => {
