+++ lib/middleware/receive-message/signup-get.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../helpers');
-const Signup = require('../../app/models/Signup.js');
+const helpers = require('../../helpers');
+const Signup = require('../../../app/models/Signup.js');
 
 module.exports = function getSignup() {
   return (req, res, next) => { // eslint-disable-line arrow-body-style
