+++ lib/middleware/receive-message/validate.js
@@ -4,7 +4,7 @@ const logger = require('winston');
 const newrelic = require('newrelic');
 const underscore = require('underscore');
 
-const helpers = require('../helpers');
+const helpers = require('../../helpers');
 
 function validateSignup(req, res) {
   if (!req.signup) {
